fixed mistyped result

+++ src/rcore.c
@@ -2952,7 +2952,7 @@ Vector2 GetWorldToScreenEx(Vector3 position, Camera camera, int width, int heigh
     }
     else if (camera.projection == CAMERA_ORTHOGRAPHIC)
     {
-        float aspect = ((double)width/(double)height);
+        double aspect = ((double)width/(double)height);
         double top = camera.fovy/2.0;
         double right = top*aspect;
